requirements.txt
fastapi
uvicorn
praw
psycopg2-binary
python-dotenv
nltk
nrclex
